+++ test/sql2json/group-by.js
@@ -296,4 +296,62 @@ describe('SQL to JSON - Group By', () => {
         json.should.deepEqual(response);
     });
 
+
+
+    it('Group with function with column name (double quotes) and constant as arguments', () => {
+        const response = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'function',
+                value: 'date_histogram',
+                alias: null,
+                arguments: [{
+                    type: 'literal',
+                    value: 'createdAt'
+                }, {
+                    type: 'string',
+                    value: '1d'
+                }]
+            }],
+        };
+
+        const obj = new Sql2json('SELECT * FROM tablename GROUP BY date_histogram("createdAt",\'1d\')');
+        const json = obj.toJSON();
+        json.should.deepEqual(response);
+    });
+
+    it('Group with function with named arguments', () => {
+        const response = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'function',
+                value: 'date_histogram',
+                alias: null,
+                arguments: [{
+                    name: 'field',
+                    type: 'literal',
+                    value: 'createdAt'
+                }, {
+                    name: 'interval',
+                    type: 'string',
+                    value: '1d'
+                }]
+            }],
+        };
+
+        const obj = new Sql2json('SELECT * FROM tablename GROUP BY date_histogram(field="createdAt",\'interval\'=\'1d\')');
+        const json = obj.toJSON();
+        json.should.deepEqual(response);
+    });
+
 });