+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - SQL2JSON - Escaping errors in queries are now detected are result in an exception
 - JSON2SQL - `literal` types are now wrapped in double quotes  when query is generated
 - JSON2SQL - `string` types are now wrapped in single quotes when query is generated  
+- Added support for named arguments in `GROUP BY` function calls